2/24/20 Tim - Got navigation down and everything is coming together
