--- App/screens/SessionHistory.js
@@ -1,141 +1,0 @@
-import React from "react";
-import { StyleSheet, Text, View, Image } from "react-native";
-import {
-  Container,
-  Header,
-  Content,
-  Footer,
-  FooterTab,
-  Button,
-  Icon,
-  Left,
-  Body,
-  Right,
-  Title
-} from "native-base";
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center"
-  },
-  text: {
-    fontSize: 18,
-    lineHeight: 24,
-    justifyContent: "center",
-    textAlign: "center",
-    position: "absolute",
-    width: 315,
-    height: 77,
-    left: 30,
-    top: 56
-  },
-  button: {
-    borderColor: "white",
-    borderWidth: 1,
-    borderRadius: 12,
-    color: "white",
-    fontSize: 24,
-    fontWeight: "bold",
-    overflow: "hidden",
-    padding: 5,
-    textAlign: "center",
-    justifyContent: "center",
-    alignItems: "center"
-  },
-  title: {
-    justifyContent: "center",
-    alignItems: "center",
-    fontSize: 30,
-    fontWeight: "bold"
-  },
-  subText: {
-    fontSize: 13,
-    justifyContent: "center",
-    textAlign: "center"
-  }
-});
-
-class SessionHistory extends React.Component {
-  render() {
-    return (
-      <Container>
-        <Header>
-          <Left>
-            <Button transparent onPress={() => this.props.navigation.goBack()}>
-              <Icon name="arrow-back" />
-            </Button>
-          </Left>
-          <Body>
-            <Title>Header</Title>
-          </Body>
-          <Right>
-            <Button transparent>
-              <Icon name="menu" />
-            </Button>
-          </Right>
-        </Header>
-        <Content>
-          <View
-            style={{
-              justifyContent: "center",
-              alignItems: "center"
-            }}
-          >
-            <Text style={styles.subText}>
-              Tap on a Day to see its stats, or select the Month to view all
-              your stats for that Month.
-            </Text>
-            <Image source={require("../Pictures/calHist.png")} />
-            <Text style={styles.title}>January</Text>
-          </View>
-
-          {/* <Button
-            onPress={() => this.props.navigation.navigate("SideBar")}
-          >
-            <Text style={styles.button}>Side Menu</Text>
-          </Button> */}
-
-          <Button
-            onPress={() => this.props.navigation.navigate("SessionCreation")}
-          >
-            <Text style={styles.button}>Session</Text>
-          </Button>
-
-          <Button onPress={() => this.props.navigation.navigate("Activities")}>
-            <Text style={styles.button}>Activities</Text>
-          </Button>
-
-          <Button
-            onPress={() => this.props.navigation.navigate("DailyHistory")}
-          >
-            <Text style={styles.button}>Daily History</Text>
-          </Button>
-        </Content>
-        <Footer>
-          <FooterTab style={{ backgroundColor: "#c2c5cc" }}>
-            <Button
-              onPress={() => this.props.navigation.navigate("SessionHistory")}
-            >
-              <Icon name="calendar" style={{ color: "#000" }} />
-            </Button>
-            <Button
-              onPress={() => this.props.navigation.navigate("ActiveSession")}
-            >
-              <Icon active name="stopwatch" style={{ color: "#000" }} />
-            </Button>
-            <Button
-              onPress={() => this.props.navigation.navigate("Activities")}
-            >
-              <Icon name="heart" style={{ color: "#000" }} />
-            </Button>
-          </FooterTab>
-        </Footer>
-      </Container>
-    );
-  }
-}
-
-export default SessionHistory;
