+++ Pedometer API/stepCounter.js
