+++ package.json
@@ -8,19 +8,22 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@react-native-community/masked-view": "0.1.5",
+    "@react-native-community/masked-view": "^0.1.6",
+    "@react-navigation/drawer": "^5.0.7",
     "expo": "~36.0.0",
-    "expo-sensors": "~8.0.0",
+    "expo-font": "^8.0.0",
     "native-base": "^2.13.8",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.0.tar.gz",
-    "react-native-gesture-handler": "~1.5.0",
-    "react-native-safe-area-context": "0.6.0",
-    "react-native-vector-icons": "^6.6.0",
+    "react-native-gesture-handler": "^1.6.0",
+    "react-native-reanimated": "^1.7.0",
+    "react-native-safe-area-context": "^0.7.3",
+    "react-native-screens": "^2.0.0-beta.10",
     "react-native-web": "~0.11.7",
-    "react-navigation": "^4.0.10",
-    "react-navigation-stack": "^2.0.15"
+    "react-navigation": "^4.1.1",
+    "react-navigation-drawer": "^2.3.4",
+    "react-navigation-stack": "^2.1.1"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
